Adapt secrets manager tool to work with TOML files and improve its structure
